+++ front-end/gdf-hrms/src/assets/gdfLogo.png
