+++ front-end/gdf-hrms/src/components/CareerHistoryLayout.js
@@ -4,8 +4,8 @@ import Paper from '@material-ui/core/Paper';
 import Grid from '@material-ui/core/Grid';
 import TextField from '@material-ui/core/TextField';
 import SimpleSelect from './SimpleSelect';
-import CareerHistoryForm  from './CareerHistoryForm';
-import CareerHistoryTable from './CareerHistoryTable';
+import CareerHistoryForm  from './CareerHistoryComponents/CareerHistoryForm';
+import CareerHistoryTable from './CareerHistoryComponents/CareerHistoryTable';
 
 
 const useStyles = makeStyles((theme) => ({
