+++ front-end/gdf-hrms/src/components/ContentArea.js
