+++ front-end/gdf-hrms/src/components/EmployeeProfileComponents/ContactForm.js
@@ -3,7 +3,7 @@ import { makeStyles } from '@material-ui/core/styles';
 import Paper from '@material-ui/core/Paper';
 import Grid from '@material-ui/core/Grid';
 import TextField from '@material-ui/core/TextField';
-import SimpleSelect from './SimpleSelect';
+import SimpleSelect from '../SimpleSelect';
 
 
 
@@ -24,7 +24,7 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-export default function OfficialInformationForm() {
+export default function ContactForm() {
   const classes = useStyles();
   const SexOption = ["Male","Female"];
   function FormRow() {
@@ -32,30 +32,24 @@ export default function OfficialInformationForm() {
       <React.Fragment>
         <div>
             <Grid item xs={2}>
-            <TextField id="RegimentalNumber" label="Home Number" variant="outlined" size="small" />
+            <TextField id="HomeNumber" label="Home Number" variant="outlined" size="small" />
             </Grid>
         </div>
         <div>
             <Grid item xs={2}>
-                <TextField id="IdentificationNumber" label="Cell Number" variant="outlined" size="small" />
+                <TextField id="CellNumber" label="Cell Number" variant="outlined" size="small" />
             </Grid>
         </div>
-        <div>
+        <div>   
             <Grid item xs={2}>
-            <TextField id="PassportNumber" label="Work Number" variant="outlined" size="small" />
+            <TextField id="WorkNumber" label="Work Number" variant="outlined" size="small" />
             </Grid >
-        </div>
-        <div>
+         </div>
+         <div>
             <Grid item xs={2}>
-            <TextField id="PassportExpirationDate" label="Passport Expiration Date" variant="outlined" size="small" InputLabelProps={{ shrink: true,}} type="date"/>
+            <TextField id="Email" label="Email" variant="outlined" size="small" />
             </Grid >
         </div>    
-        <div>
-        <Grid item xs={2}>
-            <TextField id="NISNumber" label="NIS Number" variant="outlined" size="small" />
-            </Grid>
-        </div>
-    
       </React.Fragment>
       
     );
@@ -64,7 +58,7 @@ export default function OfficialInformationForm() {
   return (
     <div className={classes.root}>
       <Grid container spacing={1}>
-      <h4>Official Information</h4>
+      <h4>Contact Information</h4>
         <Grid container item xs={12} spacing={3}>
           <FormRow />
         </Grid>
