+++ front-end/gdf-hrms/src/components/EmployeeProfileComponents/OfficialInformationForm.js
@@ -3,7 +3,7 @@ import { makeStyles } from '@material-ui/core/styles';
 import Paper from '@material-ui/core/Paper';
 import Grid from '@material-ui/core/Grid';
 import TextField from '@material-ui/core/TextField';
-import SimpleSelect from './SimpleSelect';
+import SimpleSelect from '../SimpleSelect';
 
 
 
@@ -24,7 +24,7 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-export default function ContactForm() {
+export default function OfficialInformationForm() {
   const classes = useStyles();
   const SexOption = ["Male","Female"];
   function FormRow() {
@@ -32,24 +32,30 @@ export default function ContactForm() {
       <React.Fragment>
         <div>
             <Grid item xs={2}>
-            <TextField id="HomeNumber" label="Home Number" variant="outlined" size="small" />
+            <TextField id="RegimentalNumber" label="Home Number" variant="outlined" size="small" />
             </Grid>
         </div>
         <div>
             <Grid item xs={2}>
-                <TextField id="CellNumber" label="Cell Number" variant="outlined" size="small" />
+                <TextField id="IdentificationNumber" label="Cell Number" variant="outlined" size="small" />
             </Grid>
         </div>
-        <div>   
+        <div>
             <Grid item xs={2}>
-            <TextField id="WorkNumber" label="Work Number" variant="outlined" size="small" />
+            <TextField id="PassportNumber" label="Work Number" variant="outlined" size="small" />
             </Grid >
-         </div>
-         <div>
+        </div>
+        <div>
             <Grid item xs={2}>
-            <TextField id="Email" label="Email" variant="outlined" size="small" />
+            <TextField id="PassportExpirationDate" label="Passport Expiration Date" variant="outlined" size="small" InputLabelProps={{ shrink: true,}} type="date"/>
             </Grid >
         </div>    
+        <div>
+        <Grid item xs={2}>
+            <TextField id="NISNumber" label="NIS Number" variant="outlined" size="small" />
+            </Grid>
+        </div>
+    
       </React.Fragment>
       
     );
@@ -58,7 +64,7 @@ export default function ContactForm() {
   return (
     <div className={classes.root}>
       <Grid container spacing={1}>
-      <h4>Contact Information</h4>
+      <h4>Official Information</h4>
         <Grid container item xs={12} spacing={3}>
           <FormRow />
         </Grid>
