+++ front-end/gdf-hrms/src/components/EmployeeProfileComponents/PersonalInformationForm.js
@@ -3,7 +3,7 @@ import { makeStyles } from '@material-ui/core/styles';
 import Paper from '@material-ui/core/Paper';
 import Grid from '@material-ui/core/Grid';
 import TextField from '@material-ui/core/TextField';
-import SimpleSelect from './SimpleSelect';
+import SimpleSelect from '../SimpleSelect';
